LayoutTests/http/wpt/mediarecorder/pause-recording-timeSlice.html

<!DOCTYPE html>
<html>
<head>
    <meta charset="utf-8">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
</head>
<body>
    <script>
function waitFor(duration)
{
    return new Promise((resolve) => setTimeout(resolve, duration));
}

promise_test(async (test) => {
    const stream = await navigator.mediaDevices.getUserMedia({ audio: true, video: true });

    const recorder = new MediaRecorder(stream);

    let counter = 0;
    recorder.ondataavailable = (e) => { counter++; }

    const startPromise = new Promise(resolve => recorder.onstart = resolve);
    recorder.start(100);
    await startPromise;

    setTimeout(() => recorder.pause(), 50);
    setTimeout(() => recorder.resume(), 1550);

    await waitFor(500);
    const pausedCounter = counter;

    await waitFor(500);
    assert_equals(counter, pausedCounter);

    return new Promise(resolve => recorder.ondataavailable = resolve);
}, "Pausing and resuming the recording should pause dataavailable event");
    </script>
</body>
</html>
